requirements.txt
# Base requirements
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
Pillow==12.1.0
django-jazzmin==3.0.1
python-decouple==3.8

# Database
psycopg2-binary==2.9.10

# API Documentation
drf-spectacular==0.28.0

# Security
django-ratelimit==4.1.0
django-csp==3.8

# Image processing
django-imagekit==5.0.0

# Rich text editor
django-ckeditor==6.7.1

# Tagging
django-taggit==5.0.1

# Audit logging
django-auditlog==3.0.0

# Testing
pytest-django==4.9.0
factory-boy==3.3.1
coverage==7.6.10

# Utilities
python-slugify==8.0.1
phonenumbers==8.13.50